Fase 2/Evidencias Proyecto/web/SAPPE/requirements.txt
django-crispy-forms
crispy-bootstrap4
xhtml2pdf
beautifulsoup4
reportlab
PyPDF2
googlemaps